refactor into fn
